Fix cli test
