+++ bddrest/documentary/__init__.py
@@ -7,6 +7,8 @@ from .documenter import Documenter
 
 class DocumentaryLauncher(SubCommand):
     __command__ = 'document'
+    __help__ = 'Generates REST API Documentation from standard input to ' \
+        'standard output.'
     __arguments__ = [
         Argument(
             '-f', '--format',
@@ -14,11 +16,6 @@ class DocumentaryLauncher(SubCommand):
             help='The output format. One of markdown, html. Default is '
                 'markdown.'
         ),
-        Argument(
-            'document',
-            help='Generates REST API Documentation from standard input to '
-                'standard output.'
-        )
     ]
 
     formatters = {
@@ -26,15 +23,15 @@ class DocumentaryLauncher(SubCommand):
         # 'html': HtmlFormatter,
     }
 
-    def __call__(self):
-        self.convert_file(sys.stdin, sys.stdout)
+    def __call__(self, args):
+        self.convert_file(sys.stdin, sys.stdout, args.format)
 
-    def convert_file(self, source, destination):
+    def convert_file(self, source, destination, format):
         from ..authoring import Story
         story = Story.load(source)
         story.document(
             destination,
-            formatter_factory=self.formatters[self.args.format]
+            formatter_factory=self.formatters[format]
         )
